+++ src/Hydrate.php
@@ -2,6 +2,8 @@
 
 namespace CerebralFart\Hydrate;
 
+use CerebralFart\Hydrate\Exceptions\HydrationException;
+use CerebralFart\Hydrate\Exceptions\UninitializedPropertyException;
 use ReflectionClass;
 
 class Hydrate {
@@ -10,6 +12,7 @@ class Hydrate {
      * @param $data
      * @param class-string<T> $type
      * @return T
+     * @throws HydrationException
      */
     public static function load($data, string $type): mixed {
         $refClass = new ReflectionClass($type);
@@ -20,6 +23,8 @@ class Hydrate {
             $name = $property->getName();
             if (array_key_exists($name, $data)) {
                 $property->setValue($instance, $data[$name]);
+            } else if (!$property->hasDefaultValue()) {
+                throw new UninitializedPropertyException($type, $name);
             }
         }